refactor(merge): merge ApplicationViewer components
- rename section files
- organize imports
- remove temporary logic
