+++ apps/ui/src/pages/App.tsx
@@ -21,8 +21,8 @@ import { Route, Routes } from 'react-router';
 
 import PageLayout from '@/components/layouts/PageLayout';
 import ApplicationViewer from '@/pages/applications/application';
-import Introduction from '@/pages/applications/sections/Introduction';
-import SectionA from '@/pages/applications/sections/SectionA';
+import Applicant from '@/pages/applications/sections/applicant';
+import Introduction from '@/pages/applications/sections/intro';
 import DashboardPage from '@/pages/dashboard';
 import HomePage from '@/pages/index';
 
@@ -34,11 +34,8 @@ function App() {
 				<Route path="dashboard" element={<DashboardPage />} />
 				<Route path="application/:id" element={<ApplicationViewer />}>
 					<Route path="intro/edit?" element={<Introduction />} />
-					<Route path="applicant/edit?" element={<SectionA />} />
+					<Route path="applicant/edit?" element={<Applicant />} />
 				</Route>
-				{/* TODO: merge */}
-				{/* <Route path="application/:id" element={<ApplicationIndex isEditMode={false} />} />
-				<Route path="application/:id/edit" element={<ApplicationIndex isEditMode={true} />} /> */}
 			</Route>
 		</Routes>
 	);
