+++ apps/ui/src/pages/applications/application.tsx
@@ -17,20 +17,81 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+import { Application } from '@pcgl-daco/data-model';
 import { Col, Flex, Layout, Row } from 'antd';
-import { Outlet } from 'react-router';
+import { useEffect, useState } from 'react';
+import { useTranslation } from 'react-i18next';
+import { Outlet, useMatch, useNavigate, useParams } from 'react-router';
 
-import ContentWrapper from '@/components/layouts/ContentWrapper';
+import ContentWrapper, { contentWrapperStyles } from '@/components/layouts/ContentWrapper';
 import AppHeader from '@/components/pages/application/AppHeader';
 import SectionMenu from '@/components/pages/application/SectionMenu';
+import { SkeletonLoader } from '@/components/SkeletonLoader';
+import { useGetData } from '@/global/hooks/useGetData';
+import { FetchError, ServerError } from '@/global/types';
+import { ApplicationStates } from '@pcgl-daco/data-model/dist/types';
 
 const { Content } = Layout;
 
 const ApplicationViewer = () => {
+	const params = useParams();
+	const navigation = useNavigate();
+	const { t: translate } = useTranslation();
+	const match = useMatch('/application/:id/:section/:edit');
+	const isEditMode = !!match?.params.edit;
+
+	const [applicationData, setApplicationData] = useState<Application | undefined>(undefined);
+	const [loading, setLoading] = useState<boolean>(true);
+	const [error, setError] = useState<ServerError | undefined>(undefined);
+
+	const data = useGetData(`/applications/${params.id}`) as Application | FetchError;
+
+	useEffect(() => {
+		if (data && !('isError' in data)) {
+			/**
+			 * This likely means that the user directly linked the edit page somehow
+			 * when the application was no longer in DRAFT mode. We redirect back to
+			 * the view mode as protection.
+			 *
+			 * In the future, this should also check user ability (can they edit?)
+			 */
+			if (data.state !== ApplicationStates.DRAFT && isEditMode) {
+				navigation(`/application/${data.id}/`, { replace: true });
+			}
+
+			setApplicationData(data);
+			setLoading(false);
+		} else if (data && data.isError) {
+			setLoading(false);
+
+			setError({
+				message: data.statusCode === 404 ? translate('errors.applicationNotFound.title') : data.message,
+				errors: data.statusCode === 404 ? translate('errors.applicationNotFound.message') : data.errors,
+			});
+		}
+	}, [data, isEditMode, navigation, translate]);
+
+	// Should make a component for this
+	if (!applicationData || error || loading)
+		return (
+			<Content>
+				<Row style={{ ...contentWrapperStyles }}>
+					{loading ? (
+						<SkeletonLoader />
+					) : (
+						<Col>
+							<h1>{error?.message}</h1>
+							<h2>{error?.errors}</h2>
+						</Col>
+					)}
+				</Row>
+			</Content>
+		);
+
 	return (
 		<Content>
 			<Flex style={{ height: '100%' }} vertical>
-				<AppHeader />
+				<AppHeader appId={applicationData.id} />
 				{/* Multipart form Viewer */}
 				<ContentWrapper style={{ minHeight: '70vh', padding: '2em 0', gap: '3rem' }}>
 					<>
